Move life calendar tests to own test file
